pyproject.toml
[project]
name = "security-agent"
version = "0.1.0"
description = "Security event analysis agent"
authors = [
    {name = "Varun Gurram", email = "gurramvarundeep@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "langgraph>=0.2.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "typer>=0.12.0,<1.0.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*"]

[tool.setuptools]
py-modules = ["state", "workflow", "cli"]
